+++ motohub/src/main/java/com/motohub/service/ProductService.java
@@ -14,23 +14,28 @@ public class ProductService {
     @Autowired
     private ProductRepository productRepository;
 
+    // Guardar un producto
     public Product save(Product product) {
         return productRepository.save(product);
     }
 
+    // Obtener todos los productos
     public List<Product> findAll() {
         return productRepository.findAll();
     }
 
-    public Optional<Product> findById(int id) {
+    // Buscar un producto por ID
+    public Optional<Product> findById(Long id) {
         return productRepository.findById(id);
     }
 
+    // Actualizar un producto
     public Product update(Product product) {
         return productRepository.save(product);
     }
 
-    public void delete(int id) {
+    // Eliminar un producto
+    public void delete(Long id) {
         productRepository.deleteById(id);
     }
 }
